main/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
itchat==1.3.10
lxml==4.5.2
Pillow==7.2.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
python-telegram-bot==13.1
